pyproject.toml
[project]
name = "adk-textbook"
version = "0.1.1"
description = "ADK textbook examples with agents and evaluation tools"
requires-python = ">=3.13"
dependencies = [
    # Core ADK dependencies (all chapters)
    "google-adk[eval]>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.0",
    # Chapter 7: PostgreSQL Runtime dependencies
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
    # Chapter 7: Web UI Plugin System dependencies
    "click>=8.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-fasthtml>=0.12.25",
    # Chapter 8: UI Framework dependencies
    "textual>=0.87.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    # Testing framework
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "pytest-docker>=2.0.0",
    
    # Code quality tools
    "black>=23.0.0",
    "isort>=5.12.0", 
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["textbook-adk-ch01/", "textbook-adk-ch02/", "textbook-adk-ch07-runtime/"]
markers = ["unit"]
log_level = "ERROR"
log_cli = false
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = ["ignore::UserWarning"]

# Add tool configuration for Chapter 7
[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]


[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pyright>=1.1.405",
]
